Move common webpack configs to webpack.config.js
